+++ cookbook/sample_reset_flag.py
@@ -1,6 +1,6 @@
 from triplea.service.repository import persist
 
-
 if __name__ == "__main__":
-    persist.change_status(3,2)
+    persist.change_flag_extract_topic(1,0)
+    persist.change_flag_affiliation_mining(1,0)
     persist.refresh()