+++ triplea/db/mongo_nav.py
@@ -85,25 +85,7 @@ def get_flag():
     print(list(col_article.aggregate(pipeline)))
 
 
-def change():
-    _connection_url = SETTINGS.AAA_MONGODB_CONNECTION_URL
-    client = MongoClient(_connection_url)
-    db = client[SETTINGS.AAA_MONGODB_DB_NAME]
-    col_article = db["articledata"]
-    myquery = {"FlagAffiliationMining": 1}
-    sett = {"$set": {"FlagAffiliationMining": 0}}
-    r = col_article.update_many(myquery, sett)
-    print(f"result: {r}")
 
-def change_reset_flag_affiliation_mining():
-    _connection_url = SETTINGS.AAA_MONGODB_CONNECTION_URL
-    client = MongoClient(_connection_url)
-    db = client[SETTINGS.AAA_MONGODB_DB_NAME]
-    col_article = db["articledata"]
-    myquery = {"FlagAffiliationMining": 1}
-    sett = {"$set": {"FlagAffiliationMining": 0}}
-    r = col_article.update_many(myquery, sett)
-    print(f"result: {r}")
 
 
 def change_CiteCrawlerDeep():
@@ -120,15 +102,6 @@ def change_CiteCrawlerDeep():
     print(f"result: {r}")
 
 
-def change_State():
-    _connection_url = SETTINGS.AAA_MONGODB_CONNECTION_URL
-    client = MongoClient(_connection_url)
-    db = client[SETTINGS.AAA_MONGODB_DB_NAME]
-    col_article = db["articledata"]
-    myquery = {"State": 3}
-    sett = {"$set": {"State": 2}}
-    r = col_article.update_many(myquery, sett)
-    print(f"result: {r}")
 
 
 def change_complex():