+++ triplea/service/repository/persist.py
@@ -227,6 +227,11 @@ def get_article_group_by_state():
 def change_flag_extract_topic(current_value, set_value):
     return db.change_flag_extract_topic(current_value, set_value)
 
+def change_flag_affiliation_mining(
+                                    current_value=1,
+                                    set_value=0):
+    return db.change_flag_affiliation_mining(current_value, set_value)
+
 def change_status(current_status, set_status):
     return db.change_status(current_status, set_status)